+++ Artemis/Artemis/Models/Profiles/LayerPropertiesModel.cs
@@ -11,10 +11,10 @@ namespace Artemis.Models.Profiles
     [XmlInclude(typeof(MatrixTransform))]
     public class LayerPropertiesModel
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
         public double Opacity { get; set; }
         public bool ContainedBrush { get; set; }
         public LayerAnimation Animation { get; set; }
